requirements.txt
langchain~=0.3.26
langgraph~=0.5.0
google-generativeai
chromadb
faiss-cpu
PyPDF2
python-dotenv~=1.1.1
langchain-community~=0.3.26
langchain-google-genai~=2.1.5